--- a/iree/compiler/Codegen/LLVMCPU/BUILD
+++ b/iree/compiler/Codegen/LLVMCPU/BUILD
@@ -19,7 +19,7 @@ cc_library(
         "LLVMCPUPadWorkgroupTiles.cpp",
         "LLVMCPUPlanConvLoopOrder.cpp",
         "LLVMCPUSynchronizeSymbolVisibility.cpp",
-        "LLVMCPUTileAndVectorize.cpp",
+        "LLVMCPUTileAndVectorizeLinalgTensorOps.cpp",
         "LLVMCPUUnfuseFMAOps.cpp",
         "LLVMCPUVectorization.cpp",
         "Passes.cpp",