--- a/iree/compiler/Codegen/LLVMCPU/CMakeLists.txt
+++ b/iree/compiler/Codegen/LLVMCPU/CMakeLists.txt
@@ -22,7 +22,7 @@ iree_cc_library(
     "LLVMCPUPadWorkgroupTiles.cpp"
     "LLVMCPUPlanConvLoopOrder.cpp"
     "LLVMCPUSynchronizeSymbolVisibility.cpp"
-    "LLVMCPUTileAndVectorize.cpp"
+    "LLVMCPUTileAndVectorizeLinalgTensorOps.cpp"
     "LLVMCPUUnfuseFMAOps.cpp"
     "LLVMCPUVectorization.cpp"
     "Passes.cpp"