--- a/iree/compiler/Codegen/LLVMCPU/LLVMCPULowerExecutableTarget.cpp
+++ b/iree/compiler/Codegen/LLVMCPU/LLVMCPULowerExecutableTarget.cpp
@@ -157,6 +157,9 @@ void LLVMCPULowerExecutableTargetPass::runOnOperation() {
         case IREE::HAL::DispatchLoweringPassPipeline::CPUVectorization:
           addCPUVectorizationPassPipeline(nestedModulePM, lowerToVectors);
           break;
+        case IREE::HAL::DispatchLoweringPassPipeline::CPUTensorToVectors:
+          addTensorToVectorsPassPipeline(nestedModulePM);
+          break;
         default:
           llvm_unreachable("Unsupported pipeline on CPU target.");
       }