--- a/iree/compiler/Codegen/LLVMCPU/Passes.cpp
+++ b/iree/compiler/Codegen/LLVMCPU/Passes.cpp
@@ -18,11 +18,6 @@
 namespace mlir {
 namespace iree_compiler {
 
-static llvm::cl::opt<bool> clUseTensorPadTileAndVectorize(
-    "iree-codegen-linalg-to-llvm-use-tensor-to-vectors",
-    llvm::cl::desc("If enabled will use tensor -> vector transformation pass"),
-    llvm::cl::init(false));
-
 static Value cpuAllocationFunction(OpBuilder &builder, Location loc,
                                    ArrayRef<int64_t> staticShape,
                                    Type elementType,
@@ -39,27 +34,38 @@ void addCPUVectorizationPassPipeline(OpPassManager &passManager,
   // re-enable.
   // passManager.addNestedPass<FuncOp>(createPadLinalgWorkgroupTilesPass());
 
-  if (clUseTensorPadTileAndVectorize) {
-    // Tile and vectorize linalg ops on tensors.
-    passManager.addNestedPass<FuncOp>(createLLVMCPUTilePadAndVectorizePass());
-    passManager.addNestedPass<FuncOp>(createCSEPass());
-    passManager.addNestedPass<FuncOp>(createCanonicalizerPass());
-  }
+  // Use stack allocation on CPU side.
+  addLinalgBufferizePasses(passManager, cpuAllocationFunction);
+  passManager.addNestedPass<FuncOp>(createCSEPass());
+  passManager.addNestedPass<FuncOp>(createCanonicalizerPass());
+
+  // Tile and vectorize linalg ops on buffers.
+  passManager.addNestedPass<FuncOp>(
+      createLLVMCPUVectorizationPass(lowerToVectors));
+  passManager.addNestedPass<FuncOp>(createCSEPass());
+  passManager.addNestedPass<FuncOp>(createCanonicalizerPass());
+
+  passManager.addNestedPass<FuncOp>(createForOpCanonicalizationPass());
+
+  passManager.addNestedPass<FuncOp>(createLLVMCPUPlanConvLoopOrderPass());
+}
+
+void addTensorToVectorsPassPipeline(OpPassManager &passManager) {
+  passManager.addPass(createCanonicalizerPass());
+
+  // Tile and vectorize linalg ops on tensors.
+  passManager.addNestedPass<FuncOp>(createLLVMCPUTileAndVectorizePass());
+  passManager.addNestedPass<FuncOp>(createCSEPass());
+  passManager.addNestedPass<FuncOp>(createCanonicalizerPass());
 
   // Use stack allocation on CPU side.
   addLinalgBufferizePasses(passManager, cpuAllocationFunction);
   passManager.addNestedPass<FuncOp>(createCSEPass());
   passManager.addNestedPass<FuncOp>(createCanonicalizerPass());
 
-  if (!clUseTensorPadTileAndVectorize) {
-    // Tile and vectorize linalg ops on buffers.
-    passManager.addNestedPass<FuncOp>(
-        createLLVMCPUVectorizationPass(lowerToVectors));
-    passManager.addNestedPass<FuncOp>(createCSEPass());
-    passManager.addNestedPass<FuncOp>(createCanonicalizerPass());
-  }
-
   passManager.addNestedPass<FuncOp>(createForOpCanonicalizationPass());
+
+  passManager.addNestedPass<FuncOp>(createOptimizeVectorTransferPass());
 
   passManager.addNestedPass<FuncOp>(createLLVMCPUPlanConvLoopOrderPass());
 }