--- a/iree/compiler/Codegen/Passes.h
+++ b/iree/compiler/Codegen/Passes.h
@@ -129,7 +129,8 @@ std::unique_ptr<OperationPass<ModuleOp>>
 createLLVMCPUSynchronizeSymbolVisibilityPass();
 
 /// Multi-level tiling and vectorization of linalg ops on tensors.
-std::unique_ptr<OperationPass<FuncOp>> createLLVMCPUTileAndVectorizePass();
+std::unique_ptr<OperationPass<FuncOp>> createLLVMCPUTileAndVectorizePass(
+    bool lowerToVectors = true);
 
 /// Vectorizes linalg ops executed in the same hal.interface.workgroup.
 std::unique_ptr<OperationPass<FuncOp>> createLLVMCPUVectorizationPass(
@@ -170,7 +171,8 @@ void addCPUVectorizationPassPipeline(OpPassManager &passManager,
 
 /// Populates the passes needed to multi level tile and lowering of linalg ops
 /// on tensors to vectors operations.
-void addTensorToVectorsPassPipeline(OpPassManager &passManager);
+void addTensorToVectorsPassPipeline(OpPassManager &passManager,
+                                    bool lowerToVectors = true);
 
 //----------------------------------------------------------------------------//
 // LLVMCPU Pass Pipelines for lowering to LLVM dialect.