--- a/iree/compiler/Codegen/Passes.td
+++ b/iree/compiler/Codegen/Passes.td
@@ -123,11 +123,10 @@ def LLVMCPUSynchronizeSymbolVisibility :
   let constructor = "mlir::iree_compiler::createLLVMCPUSynchronizeSymbolVisibilityPass()";
 }
 
-def LLVMCPUTilePadAndVectorize :
-    Pass<"iree-llvmcpu-tile-pad-and-vectorize", "FuncOp"> {
-  let summary = "Tile and pad workgroups tiles";
-  let constructor =
-      "mlir::iree_compiler::createLLVMCPUTilePadAndVectorizePass()";
+def LLVMCPUTileAndVectorize
+    : Pass<"iree-llvmcpu-tile-and-vectorize", "FuncOp"> {
+  let summary = "Tile and vectorize workgroups tiles";
+  let constructor = "mlir::iree_compiler::createLLVMCPUTileAndVectorizePass()";
 }
 
 def LLVMCPUUnfuseFMAOps :