@asaadaldien 2021-09-14
I incorrectly parsed the log `integrations/tensorflow/bindings/python/iree/tf/support/tf_test_utils_test.py` is the failing test not `MobileNetV2`, Its strange since there is nothing special about the matmul in that test!