@MaheshRavishankar 2021-09-15
Some of the benchmark models are failing too. Those are easier to triage. I think the steps are you have to set `-DIREE_BUILD_BENCHMARKS` to true and do `cmake --build . --target benchmark_suites`. Then you should see `benchmark_suites` folder in the build directory (I might have the order wrong, you might not need to build). The `CMakeLists` in `iree/benchmarks/TensorFlow/` should have the options to use.