@asaadaldien 2021-09-15
> This is fine, but there might be a better way. You can look at the `hal.executable.variant` to see if this is vmvx or not and then vectorize or not. Then we can drop all of these flags. Fine to do this in a later pass (let me know if you are going to do it in a later pass)

Sure, I can do this in second pass, but not sure which flags you are looking to drop?