@MaheshRavishankar 2021-09-15
> Sure, I can do this in second pass, but not sure which flags you are looking to drop?

This flag here (https://github.com/google/iree/blob/fd7dd286d8cc653b27f85b9bf07a9683d9bf7abe/iree/compiler/Codegen/LLVMCPU/LLVMCPULowerExecutableTarget.cpp#L75). This is set when called from the executable translation. Instead, just look for the backend within the pass itself to figure it out so that all that plumbing can be cleaned up.