@asaadaldien 2021-09-16
> This PR seems to cause the CI failure at ToT.
> 
> https://source.cloud.google.com/results/invocations/6112d1b9-8962-4bb0-8bad-16e1ded18ec8/targets/iree%2Fgcp_ubuntu%2Fbazel%2Flinux%2Fx86-swiftshader%2Fcore%2Fmain/log

Thanks for the heads up, seems like a canonicalized was merged upstream caused this. Should be fixed by #7065